% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_cal.R
\name{finalize_cal}
\alias{finalize_cal}
\title{Finish: Final Processing Steps for Calendar Data}
\usage{
finalize_cal(dat, write = TRUE, path = NULL)
}
\arguments{
\item{dat}{A data file that has been passed to \emph{all} of the vcal
functions in this package.}

\item{write}{Logical: write file as compressed CSV? (defaults TRUE)}

\item{path}{Optional Character: path to the folder "unlinked_lr_data". If
not provided and write == TRUE, a file will be written to \code{
general/calendar/unlinked_lr_data} in the \code{dhs} folder.}
}
\description{
Run this function after you have completed all of the vcal
functions in this package (a check will be run to ensure that all of the
expected variables are present, even if all values are NA). Returns a data
frame and writes a compressed CSV file by default.
}
\author{
Matt Gunther
}
