% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcal_bfeed.R
\name{vcal_bfeed}
\alias{vcal_bfeed}
\title{Breastfeeding Calendar}
\usage{
vcal_bfeed(dat)
}
\arguments{
\item{dat}{A data file created by \code{vcal_reprod()} (may be passed to
any other function starting with "vcal" first).}
}
\description{
Create all variables related to the Breastfeeding calendar
(see details). If this calendar was not included in the sample (or if some
of the required information is not available), all variables will still be
 created, but all values will be NA.
}
\details{
The following variables will be created using case logic provided
to the function \code{dplyr::case_when()}. Please note that
\code{case_when()} returns NA through implicit logic: \emph{if a "case"
exists and is not explicitly handled here, the value NA will be returned!}
\itemize{
  \item{
    \strong{breastf} Integer: may take one of three values, and is not
    available (NA) otherwise:
      \itemize{
        \item{1 - \code{vcal_bfeed} contains "X" or "x"}
        \item{2 - \code{vcal_bfeed} contains "N" or "n"}
        \item{0 - \code{vcal_bfeed} is 0}
      }
  }
}
}
\author{
Matt Gunther
}
