% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcal_marstat.R
\name{vcal_marstat}
\alias{vcal_marstat}
\title{Marital Status Calendar}
\usage{
vcal_marstat(dat)
}
\arguments{
\item{dat}{A data file created by \code{vcal_reprod()} (may be passed to
any other function starting with "vcal" first).}
}
\description{
Create all variables related to the Marital Status calendar
(see details). If this calendar was not included in the sample (or if some
of the required information is not available), all variables will still be
 created, but all values will be NA.
}
\details{
The following variables will be created using case logic provided
to the function \code{dplyr::case_when()}. Please note that
\code{case_when()} returns NA through implicit logic: \emph{if a "case"
exists and is not explicitly handled here, the value NA will be returned!}
\itemize{
  \item{
    \strong{married} Logical: TRUE if \code{vcal_marstat} contains "X" or
    "x". FALSE if  \code{vcal_marstat} is 0. Not available if vcal_marstat
    is not available (NA) or any other character (e.g. " " or "?").
  }
}
}
\author{
Matt Gunther
}
