% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcal_work.R
\name{vcal_work}
\alias{vcal_work}
\title{Employment Calendar}
\usage{
vcal_work(dat)
}
\arguments{
\item{dat}{A data file created by \code{vcal_reprod()} (may be passed to
any other function starting with "vcal" first).}
}
\description{
Create all variables related to the Employment calendar
(see details). If this calendar was not included in the sample (or if some
of the required information is not available), all variables will still be
 created, but all values will be NA.
}
\details{
The following variables will be created using case logic provided
to the function \code{dplyr::case_when()}. Please note that
\code{case_when()} returns NA through implicit logic: \emph{if a "case"
exists and is not explicitly handled here, the value NA will be returned!}
\itemize{
  \item{
    \strong{empl} Integer: records the numeric value in \code{vcal_work}.
    Not available if \code{vcal_work} is missing (" "), not available (NA),
    or any non-numeric value.
  }
}
}
\author{
Matt Gunther
}
